loan calculator part-3 - loader user experience
Added loader, due to which a gif will load indicating that the calculation is being processed. The form was not accepting and processing floating-point values as the condition 'step=any' wasn't declared in the input field for Interest. Added 'step=any' field and now the form is accepting decimal values and processing them as well.